# lepiossa-golang-2022-backend-dini-surya/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/os": {
            "get": {
                "description": "Get a list of OperasiSistem.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperasiSistem"
                ],
                "summary": "Get all OperasiSistem.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OperasiSistem"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new OperasiSistem.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperasiSistem"
                ],
                "summary": "Create New OperasiSistem.",
                "parameters": [
                    {
                        "description": "the body to create a new OperasiSistem",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.OperasiSistemInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OperasiSistem"
                        }
                    }
                }
            }
        },
        "/os/{id}": {
            "get": {
                "description": "Get an OperasiSistem by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperasiSistem"
                ],
                "summary": "Get OperasiSistem.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OperasiSistem id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OperasiSistem"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a OperasiSistem by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperasiSistem"
                ],
                "summary": "Delete one OperasiSistem.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OperasiSistem id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update OperasiSistem by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperasiSistem"
                ],
                "summary": "Update OperasiSistem.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OperasiSistem id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update age os category",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.OperasiSistemInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OperasiSistem"
                        }
                    }
                }
            }
        },
        "/os/{id}/laptop": {
            "get": {
                "description": "Get all Laptops by OperasiSistemId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperasiSistem"
                ],
                "summary": "Get Laptops.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OperasiSistem id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Laptop"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.OperasiSistemInput": {
            "type": "object",
            "properties": {
                "namaOs": {
                    "type": "string"
                }
            }
        },
...
        "models.Laptop": {
            "type": "object",
            "properties": {
                "brandID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "grafis": {
                    "type": "string"
                },
                "hardisk": {
                    "type": "string"
                },
                "harga": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "layar": {
                    "type": "string"
                },
                "namaLaptop": {
                    "type": "string"
                },
                "osID": {
                    "type": "integer"
                },
                "prosesor": {
                    "type": "string"
                },
                "ram": {
                    "type": "string"
                },
                "ratingID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.OperasiSistem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "namaOs": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}